+++ src/components/BlogForm/BuilderSearch.tsx
@@ -71,6 +71,7 @@ const BuilderSearch = () => {
         ))}
       </Menu>
       {addingBlock === "Text" && <TextModal active={true} />}
+      {/* {true && <TextModal active={true} />} */}
     </Container>
   );
 };
